chore: show entity configuration json  (#155)
* feat: tools validator metadata

* feat: test tools metadata

* fix: removed imports unused

* feat: tools authn request spid/cie

* feat: test tools authn request

* fix: code cleaned up with bash linting

* feat: tool for validate entity configuration

* fix: handle field name for validation page

* chore: show entity configuration json after successfully validation

* fix: title of page create jwk

* feat: User login session history with local logout

* chore: handled the EC validator exception

* fix: remove js unused

* fix: removed link

* fix: show json of  Entity configuration

+++ spid_cie_oidc/provider/models.py
@@ -41,9 +41,9 @@ class OidcSession(TimeStampedModel):
         except Exception:
             logger.warning(f"Error setting SID for OidcSession {self}")
 
-    def revoke(self):
+    def revoke(self, destroy_session=True):
         session = Session.objects.filter(session_key=self.sid)
-        if session:
+        if session and destroy_session:
             session.delete()
         self.revoked = True
         iss_tokens = IssuedToken.objects.filter(session=self)
@@ -67,6 +67,7 @@ class OidcSession(TimeStampedModel):
         verbose_name = "User Session"
         verbose_name_plural = "User Sessions"
         unique_together = ("client_id", "nonce")
+        ordering = ["-created"]
 
 
 class IssuedToken(TimeStampedModel):
